organizacao para testes
