+++ .gitignore
@@ -7,3 +7,4 @@ dist/
 .C4Z
 test_files/OLD/
 coboljsonifier.egg-info/
+.vscode
