+++ src/tests/extractors/test_field_extractor.py
@@ -1,7 +1,7 @@
 import unittest
 
-from src.coboljsonifier.extractors.book_item import BookItem
-from src.coboljsonifier.extractors.field_extractor import   FieldEmpty, FieldSimpleNumeric1Decimals2, FieldSimpleNumericDecimals1, FieldSimpleNumericDecimals2, \
+from coboljsonifier.extractors.book_item import BookItem
+from coboljsonifier.extractors.field_extractor import   FieldEmpty, FieldSimpleNumeric1Decimals2, FieldSimpleNumericDecimals1, FieldSimpleNumericDecimals2, \
                                                         FieldSimpleNumeric1Decimals2, FieldSignalNumeric1Decimals1, FieldSignalNumeric1Decimals2, \
                                                         FieldNumericMasked1, FieldAlphabetic, FieldAlphanumeric, FieldUndefined
