+++ src/tests/fields/test_numeric_comp3.py
@@ -1,8 +1,8 @@
-from src.coboljsonifier.fields.field import Field
+from coboljsonifier.fields.field import Field
 import unittest
 from decimal import Decimal
 
-from  src.coboljsonifier.fields.field_numeric_comp3 import FieldNumericComp3
+from coboljsonifier.fields.field_numeric_comp3 import FieldNumericComp3
 
 
 class TestFieldNumericComp3(unittest.TestCase):
